going to bed Thursday

+++ client/src/components/RecipeCard/RecipeCard.js
@@ -30,22 +30,30 @@ function RecipeCard({
         onClick={viewRecipe}
       />
       <div className="recipe-card-bottom">
-      <p className="recipe-description">{description}</p>
-      {recipe.user_id === user?.id && (
-        <button className="update-recipe-button" onClick={handleClick}>
-          Update Recipe
-        </button>
-      )}
-      {recipe.user_id === user?.id && (
-        <button
-          className="delete-recipe-button"
-          onClick={(e) => {
-            handleDeleteRecipe(recipe.id);
-          }}
-        >
-          Delete Recipe
-        </button>
-      )}
+        <p className="recipe-description">{description}</p>
+        {recipe.user_id === user?.id ? (
+          <button className="update-recipe-button" onClick={handleClick}>
+            Update Recipe
+          </button>
+        ) : (
+          <p>Recipe From:</p>
+        )}
+        {recipe.user_id === user?.id ? (
+          <button
+            className="delete-recipe-button"
+            onClick={(e) => {
+              if (
+                window.confirm("Are You Sure You Want To Delete This Recipe?")
+              ) {
+                handleDeleteRecipe(recipe.id);
+              }
+            }}
+          >
+            Delete Recipe
+          </button>
+        ) : (
+          <p>{recipe.user_id}</p>
+        )}
       </div>
     </li>
   );
